requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.2
python-dotenv==1.0.1
Flask-Migrate==4.0.7
Werkzeug==3.0.3
Pillow==10.4.0
requests==2.32.3
itsdangerous==2.2.0
email-validator==2.1.1
python-decouple==3.8
pytest==8.3.2
pytest-flask==1.3.0
gunicorn==22.0.0
bcrypt==4.2.0
Flask-Limiter==3.7.0
Authlib==1.3.1
google-generativeai==0.8.2
psycopg2-binary==2.9.9
# Dev
black==24.8.0
isort==5.13.2

